Edit CrazyTestersTest with BaseTest extension Add testSubmitForm
