+++ src/test/java/school/redrover/CrazyTestersTest.java
@@ -1,16 +1,18 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.interactions.Actions;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+import school.redrover.runner.BaseTest;
 
 import java.time.Duration;
 
-public class CrazyTestersTest {
+public class CrazyTestersTest extends BaseTest {
 
     @Test
     public void testDragAndDrop() {
@@ -27,5 +29,33 @@ public class CrazyTestersTest {
         Assert.assertEquals(columnB.getText(), "A");
 
         driver.quit();
+    }
+
+    @Test
+    public void testSubmitForm() {
+        WebDriver driver = getDriver();
+        JavascriptExecutor js = (JavascriptExecutor) driver;
+        driver.get("https://demoqa.com/automation-practice-form");
+
+        WebElement textName = driver.findElement(By.id("firstName"));
+        textName.sendKeys("Anastasia");
+
+        WebElement textLastName = driver.findElement(By.id("lastName"));
+        textLastName.sendKeys("Belugina");
+
+        WebElement radioButton = driver.findElement(By.xpath("//label[text()='Female']"));
+        radioButton.click();
+
+        WebElement textMobile = driver.findElement(By.id("userNumber"));
+        textMobile.sendKeys("1234567890");
+
+        WebElement submitButton = driver.findElement(By.id("submit"));
+        js.executeScript("arguments[0].scrollIntoView();", submitButton);
+        submitButton.click();
+
+        WebElement result = driver.findElement(By.id("example-modal-sizes-title-lg"));
+        String resultText = result.getText();
+
+        Assert.assertEquals("Thanks for submitting the form", resultText);
     }
 }
